inv-ring: merge draw.c
