+++ src/tr1/game/inventory_ring.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "game/inventory_ring/control.h"
-#include "game/inventory_ring/draw.h"
 #include "game/inventory_ring/vars.h"
 
 #include <libtrx/game/inventory_ring.h>
