--- src/tr1/game/inventory_ring/draw.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <libtrx/game/inventory_ring/draw.h>
