+++ src/tr2/game/inventory_ring.h
@@ -1,5 +1,4 @@
 #pragma once
 
 #include "game/inventory_ring/control.h"
-#include "game/inventory_ring/draw.h"
 #include "game/inventory_ring/vars.h"
