--- src/tr2/game/inventory_ring/draw.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <libtrx/game/inventory_ring/types.h>
-
-void InvRing_Draw(INV_RING *ring);
